pyproject.toml
[tool.poetry]
name = "ndastro-api"
version = "0.1.0"
description = "API to serve data for ndastro-ui"
authors = ["Jaganathan, Bantheswaran <jaganathan.eswaran@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ndastro_api" }
]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.13.1"
fastapi = {extras = ["standard"], version = "^0.115.13"}
alembic = "^1.16.2"
sentry-sdk = {extras = ["fastapi"], version = "^2.30.0"}
pydantic = "^2.11.7"
pydantic-settings = "^2.10.0"
pyjwt = "^2.10.1"
sqlmodel = "^0.0.24"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
emails = "^0.6"
tenacity = "^9.1.2"
python-dotenv = "^1.1.0"
bcrypt = "^4.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
skyfield = "^1.53"
pytz = "^2025.2"
gunicorn = "^23.0.0"
asyncpg = "^0.30.0"
fastcrud = "^0.15.12"
aiosqlite = "^0.21.0"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
types-passlib = "^1.7.7.20250602"
ruff = "^0.12.0"
mypy = "^1.16.1"
coverage = "^7.9.1"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
pre-start = "ndastro_api.pre_start:main"
db-migrate = "ndastro_api.db_migrator:run_migration"
init-data = "ndastro_api.initial_data:main"
test-prestart = "ndastro_api.tests_pre_start:main"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore=["INP001", "COM812", "D213", "D203"]

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["S101", "D103", "D100", "ANN201"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
env_files = ["ndastro_api/.env.test"]
testpaths = [
    "ndastro_api/tests"
]
